rename 04-react-router to 06-public-protected-routing...
The rationale is that jumping from a basic app to a complex routing
example felt like it was an overload of too many concepts to grok at one
time.

The plan is to introduce 2 more starters that will first explore css
scoping and a simple router  before getting into a full blown example
of public and protected routing.