all: move 'code' to 'src' folder...
Keeping 'code' at the same level where 'node_modules' and 'build' folder
gets created was a bad idea. Especially when we created #06 and tried to
organize the app using 'seperation of concerns' and 'testability in
isolation' principles, it was evident that we needed yet another
container to keep logical things together.